requirements.txt
# Core LangGraph and LangChain dependencies
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-google-genai>=0.8.0
langchain-community>=0.2.0

# PDF processing
pypdf2>=3.0.1
pdfplumber>=0.11.0
pymupdf>=1.24.0  # Alternative PDF processor

# Configuration and environment
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0.0

pandas>=2.0.0
numpy>=1.24.0